+++ lib/src/dto/request/payment_page/payment_page_initialize_request_params.dart
@@ -1,6 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:payyo_sdk/src/dto/common/common.dart';
 import 'package:payyo_sdk/src/dto/common/customer.dart';
+import 'package:payyo_sdk/src/dto/common/metadata.dart';
 import 'package:payyo_sdk/src/dto/common/payment_plan.dart';
 import 'package:payyo_sdk/src/dto/common/return_urls.dart';
 import 'package:payyo_sdk/src/dto/common/web_hook.dart';
@@ -14,7 +15,8 @@ class PaymentPageInitializeRequestParams extends RequestParams {
   String merchant_reference;
   String description;
   String billing_descriptor;
-  dynamic metadata;
+  @JsonKey(fromJson: metadataFromJson, toJson: metadataToJson)
+  Map<String, dynamic> metadata;
   String currency;
   int amount;
   List<Customer> customers;