+++ lib/src/dto/request/transaction/transaction_initiate_params.g.dart
@@ -13,7 +13,7 @@ TransactionInitiateParams _$TransactionInitiateParamsFromJson(
     ..merchant_reference = json['merchant_reference'] as String
     ..description = json['description'] as String
     ..billing_descriptor = json['billing_descriptor'] as String
-    ..metadata = json['metadata']
+    ..metadata = metadataFromJson(json['metadata'] as String)
     ..currency = json['currency'] as String
     ..amount = json['amount'] as int
     ..was_dcc_offered = json['was_dcc_offered'] as bool
@@ -58,7 +58,7 @@ Map<String, dynamic> _$TransactionInitiateParamsToJson(
   writeNotNull('merchant_reference', instance.merchant_reference);
   writeNotNull('description', instance.description);
   writeNotNull('billing_descriptor', instance.billing_descriptor);
-  writeNotNull('metadata', instance.metadata);
+  writeNotNull('metadata', metadataToJson(instance.metadata));
   writeNotNull('currency', instance.currency);
   writeNotNull('amount', instance.amount);
   writeNotNull('was_dcc_offered', instance.was_dcc_offered);